Plans for new functions to create the AI
Design for the board might need to be changed since implementation might be pretty challenging with the way it is. I think it might be better to either create objects, or some sort of array which keeps the position each of the pieces, which makes generating all possible moves for a team much easier (this is just my opinion though).